Remove transcripts doc page, redirect path to Captions doc.
A few months ago, we (sneakily) dropped transcript support for many accounts,
and instead switched them to captions only. This makes the existing transcripts
doc a bit confusing.

For now, this will remove the transcripts doc entirely. If lots of older
customers who have access to the transcript capabilities have lots of questions
in the future, we can bring it back.